+++ internal/generator/usecase/general/locale/consts.go
@@ -3,8 +3,11 @@ package locale
 // Common string charsets.
 
 var (
-	Numbers      = []rune("0123456789")
-	SpecialChars = []rune("!#$%&()*+,-.:;<=>?@_{|}")
+	Numbers          = []rune("0123456789")
+	SpecialChars     = []rune("!#$%&()*+,-.:;<=>?@_{|}")
+	Base64Charset    = []rune("+/0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz")
+	Base64URLCharset = []rune("-_0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz")
+	HexCharset       = []rune("0123456789ABCDEF")
 )
 
 // Genders declaration.